+++ src/blocks.js
@@ -27,6 +27,7 @@ import './blocks/form-button/index.js';
 import './blocks/form-calculation/index.js';
 import './blocks/form-column/index.js';
 import './blocks/form-group/index.js';
+import './blocks/form-step/index.js';
 import './blocks/message/index.js';
 import './blocks/name/index.js';
 import './blocks/number/index.js';
