+++ src/blocks/form-step/index.js
@@ -1,0 +1,19 @@
+import stepFormEdit from "./edit";
+import stepFormSave from "./save";
+
+registerBlockType("cwp/form-step", {
+	title: __("Form Step"),
+	icon: "editor-table",
+	category: "common",
+	keywords: [
+		__("gutenberg-forms"),
+		__("forms"),
+		__("form-step"),
+		__("step"),
+		__("multistep")
+	],
+	edit: stepFormEdit,
+	save: stepFormSave,
+	attributes: {},
+	parent: ["cwp/block-gutenberg-forms"]
+});
